# [Compiler] Produce an un-encoded program by the compiler