--- a/bbq/compiler/codegen.go
+++ b/bbq/compiler/codegen.go
@@ -19,15 +19,18 @@
 package compiler
 
 import (
+	"github.com/onflow/cadence/bbq"
 	"github.com/onflow/cadence/bbq/opcode"
 	"github.com/onflow/cadence/errors"
+	"github.com/onflow/cadence/interpreter"
 )
 
-type CodeGen[E any] interface {
+type CodeGen[E, T any] interface {
 	Offset() int
 	SetTarget(code *[]E)
 	Emit(instruction opcode.Instruction)
 	PatchJump(offset int, newTarget uint16)
+	AddType(staticType bbq.StaticType) T
 }
 
 // ByteCodeGen is a CodeGen implementation that emits bytecode
@@ -35,7 +38,7 @@ type ByteCodeGen struct {
 	target *[]byte
 }
 
-var _ CodeGen[byte] = &ByteCodeGen{}
+var _ CodeGen[byte, []byte] = &ByteCodeGen{}
 
 func (g *ByteCodeGen) Offset() int {
 	return len(*g.target)
@@ -53,12 +56,20 @@ func (g *ByteCodeGen) PatchJump(offset int, newTarget uint16) {
 	opcode.PatchJump(g.target, offset, newTarget)
 }
 
+func (g *ByteCodeGen) AddType(staticType bbq.StaticType) []byte {
+	bytes, err := interpreter.StaticTypeToBytes(staticType)
+	if err != nil {
+		panic(err)
+	}
+	return bytes
+}
+
 // InstructionCodeGen is a CodeGen implementation that emits opcode.Instruction
 type InstructionCodeGen struct {
 	target *[]opcode.Instruction
 }
 
-var _ CodeGen[opcode.Instruction] = &InstructionCodeGen{}
+var _ CodeGen[opcode.Instruction, bbq.StaticType] = &InstructionCodeGen{}
 
 func (g *InstructionCodeGen) Offset() int {
 	return len(*g.target)
@@ -89,4 +100,8 @@ func (g *InstructionCodeGen) PatchJump(offset int, newTarget uint16) {
 	default:
 		panic(errors.NewUnreachableError())
 	}
+}
+
+func (g *InstructionCodeGen) AddType(staticType bbq.StaticType) bbq.StaticType {
+	return staticType
 }