--- a/bbq/vm/linker.go
+++ b/bbq/vm/linker.go
@@ -21,7 +21,6 @@ package vm
 import (
 	"fmt"
 	"github.com/onflow/cadence/bbq"
-	"github.com/onflow/cadence/bbq/opcode"
 	"github.com/onflow/cadence/common"
 )
 
@@ -36,7 +35,7 @@ type LinkedGlobals struct {
 // LinkGlobals performs the linking of global functions and variables for a given program.
 func LinkGlobals(
 	location common.Location,
-	program *bbq.Program[opcode.Instruction, bbq.StaticType],
+	program *bbq.InstructionProgram,
 	conf *Config,
 	linkedGlobalsCache map[common.Location]LinkedGlobals,
 ) LinkedGlobals {