--- a/bbq/vm/linker.go
+++ b/bbq/vm/linker.go
@@ -20,6 +20,7 @@ package vm
 
 import (
 	"fmt"
+
 	"github.com/onflow/cadence/bbq"
 	"github.com/onflow/cadence/common"
 )