--- a/bbq/vm/native_functions.go
+++ b/bbq/vm/native_functions.go
@@ -20,6 +20,7 @@ package vm
 
 import (
 	"fmt"
+	"github.com/onflow/cadence/bbq"
 
 	"github.com/onflow/cadence/bbq/commons"
 
@@ -59,7 +60,7 @@ func RegisterTypeBoundFunction(typeName, functionName string, functionValue Nati
 func init() {
 	RegisterFunction(commons.LogFunctionName, NativeFunctionValue{
 		ParameterCount: len(stdlib.LogFunctionType.Parameters),
-		Function: func(config *Config, typeArguments []StaticType, arguments ...Value) Value {
+		Function: func(config *Config, typeArguments []bbq.StaticType, arguments ...Value) Value {
 			// TODO: Properly implement
 			fmt.Println(arguments[0].String())
 			return VoidValue{}
@@ -68,7 +69,7 @@ func init() {
 
 	RegisterFunction(commons.PanicFunctionName, NativeFunctionValue{
 		ParameterCount: len(stdlib.PanicFunctionType.Parameters),
-		Function: func(config *Config, typeArguments []StaticType, arguments ...Value) Value {
+		Function: func(config *Config, typeArguments []bbq.StaticType, arguments ...Value) Value {
 			// TODO: Properly implement
 			messageValue, ok := arguments[0].(StringValue)
 			if !ok {
@@ -84,7 +85,7 @@ func init() {
 
 	RegisterFunction(commons.GetAccountFunctionName, NativeFunctionValue{
 		ParameterCount: len(stdlib.PanicFunctionType.Parameters),
-		Function: func(config *Config, typeArguments []StaticType, arguments ...Value) Value {
+		Function: func(config *Config, typeArguments []bbq.StaticType, arguments ...Value) Value {
 			address := arguments[0].(AddressValue)
 			return NewAccountReferenceValue(config, common.Address(address))
 		},