--- a/bbq/vm/value.go
+++ b/bbq/vm/value.go
@@ -20,6 +20,7 @@ package vm
 
 import (
 	"github.com/onflow/atree"
+
 	"github.com/onflow/cadence/bbq"
 	"github.com/onflow/cadence/interpreter"
 )