--- a/bbq/vm/value_ephemeral_reference.go
+++ b/bbq/vm/value_ephemeral_reference.go
@@ -20,6 +20,7 @@ package vm
 
 import (
 	"github.com/onflow/atree"
+
 	"github.com/onflow/cadence/bbq"
 	"github.com/onflow/cadence/format"
 	"github.com/onflow/cadence/interpreter"