… bbq/vm/test/ft_test.go
… bbq/vm/test/interpreter_test.go
… bbq/vm/test/runtime_test.go
… bbq/vm/test/utils.go
… bbq/vm/test/vm_bench_test.go
… bbq/vm/test/vm_test.go
… bbq/vm/value_composite.go
… bbq/vm/value_dictionary.go
… bbq/vm/value_function.go
… bbq/vm/value_int.go
… bbq/vm/value_nil.go
… bbq/vm/value_path.go
… bbq/vm/value_simple_composite.go
… bbq/vm/value_some.go
… bbq/vm/value_storage_reference.go
… bbq/vm/value_string.go
… bbq/vm/value_ufix64.go
… bbq/vm/value_void.go
… bbq/vm/vm.go
… bbq/vm/vm_test.go