Work towards https://github.com/onflow/cadence/issues/3769

## Description

Currently, the compiler produces a program where all types are encoded (converted to bytes), and in the VM, these types are decoded back again. However, given compilation happens on the same place, this encoding and then decoding of types is an unnecessary overhead.

Therefore, updated the compiler such that it produce the decoded type, and the VM can use it directly without having to decode. This is similar to emitting instructions vs bytecode.

Micro-benchmarks showed that this reduced about 17% of the compile+execution time for FT transfer.
```
             │ ft_existing │            ft_optimized            │
             │   sec/op    │   sec/op     vs base               │
FTTransfer-8   88.97µ ± 3%   73.54µ ± 2%  -17.34% (p=0.002 n=6)

             │ ft_existing  │            ft_optimized             │
             │     B/op     │     B/op      vs base               │
FTTransfer-8   65.76Ki ± 0%   55.75Ki ± 0%  -15.22% (p=0.002 n=6)

             │ ft_existing │            ft_optimized            │
             │  allocs/op  │  allocs/op   vs base               │
FTTransfer-8   1.268k ± 0%   1.033k ± 0%  -18.53% (p=0.002 n=6)
```

We could potentially do the same for constants as well.

______

<!-- Complete: -->

- [ ] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [ ] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels